Refactor database schema and wallet handling for consistency and migration support

+++ src/Points/Wallet.php
@@ -5,9 +5,9 @@ defined( 'ABSPATH' ) || exit;
 
 class Wallet {
 	public int $user_id;
-	public float $points;
+	public int $points;
 
-	public function __construct( int $user_id = 0, float $points = 0.0 ) {
+	public function __construct( int $user_id = 0, int $points = 0 ) {
 		$this->user_id = $user_id;
 		$this->points  = $points;
 	}
